+++ src/AudioTrackManager.ts
@@ -14,16 +14,22 @@ declare const window: WindowWithAudioContext;
 class AudioTrackManager {
   private audioContext: AudioContext | undefined;
 
+  private muteForClientGainNode: GainNode | null = null;
+
+  private clientVolumeGainNode: GainNode | null = null;
+
   public isPlaying: boolean;
+
+  private clientSetVolume = 1;
+
+  private clientSetMuted = false;
 
   private currentAudio: {
     source: AudioBufferSourceNode;
-    gainNode: GainNode;
+    originalGainNode: GainNode;
     url: string;
     originalVolume: number;
   }[];
-
-  private muted = false;
 
   constructor() {
     this.isPlaying = false;
@@ -44,25 +50,30 @@ class AudioTrackManager {
     const audioBuffer = await this.loadAudioBuffer(audioTrack.url);
     if (this.audioContext === undefined) return;
 
-    const gainNode = this.audioContext.createGain();
-    gainNode.gain.value = audioTrack.volume;
+    const sourceGainNode = this.audioContext.createGain();
+    sourceGainNode.gain.value = audioTrack.volume;
 
     const source = this.audioContext.createBufferSource();
     if (audioBuffer === undefined) return;
     source.buffer = audioBuffer;
     source.loop = audioTrack.loop;
-    source.connect(gainNode).connect(this.audioContext.destination);
+    // eslint-disable-next-line @typescript-eslint/no-non-null-assertion
+    source.connect(sourceGainNode).connect(this.clientVolumeGainNode!);
     source.start(0);
 
     source.onended = () => {
       this.currentAudio = this.currentAudio.filter(
         (currentAudio) => currentAudio.source !== source,
       );
+
+      if (this.currentAudio.length === 0) {
+        this.isPlaying = false;
+      }
     };
 
     this.currentAudio.push({
       source,
-      gainNode,
+      originalGainNode: sourceGainNode,
       url: audioTrack.url,
       originalVolume: audioTrack.volume,
     });
@@ -80,12 +91,28 @@ class AudioTrackManager {
     }
 
     this.audioContext = new AudioContextClass();
+    this.clientVolumeGainNode = this.audioContext.createGain();
+    this.muteForClientGainNode = this.audioContext.createGain();
+
+    this.muteForClientGainNode.gain.setValueAtTime(
+      this.clientSetMuted ? 0 : 1,
+      this.audioContext.currentTime,
+    );
+    this.clientVolumeGainNode.gain.setValueAtTime(
+      this.clientSetVolume,
+      this.audioContext.currentTime,
+    );
+
+    this.clientVolumeGainNode.connect(this.muteForClientGainNode);
+    this.muteForClientGainNode.connect(this.audioContext.destination);
 
     return this.audioContext;
   };
 
   public async play(audioTracks: AudioTrack[]): Promise<void> {
-    if (audioTracks.length === 0) return;
+    if (audioTracks.length === 0) {
+      return;
+    }
     if (this.audioContext === undefined) {
       this.getAudioContext();
     }
@@ -121,6 +148,10 @@ class AudioTrackManager {
         }
       }),
     );
+
+    if (this.currentAudio.length === 0) {
+      this.isPlaying = false;
+    }
   }
 
   public pause(): void {
@@ -139,19 +170,39 @@ class AudioTrackManager {
   }
 
   public toggleMute(): void {
-    this.muted = !this.muted;
+    this.clientSetMuted = !this.clientSetMuted;
 
-    this.currentAudio.forEach(({ gainNode }) => {
-      // eslint-disable-next-line no-param-reassign
-      gainNode.gain.value = this.muted ? 0 : 1;
-    });
+    if (!this.audioContext || !this.muteForClientGainNode) {
+      return;
+    }
+    this.muteForClientGainNode.gain.setValueAtTime(
+      this.clientSetMuted ? 0 : 1,
+      this.audioContext.currentTime,
+    );
+  }
+
+  public setMuted(muted: boolean): void {
+    this.clientSetMuted = muted;
+    if (!this.audioContext || !this.muteForClientGainNode) {
+      return;
+    }
+    this.muteForClientGainNode.gain.setValueAtTime(
+      this.clientSetMuted ? 0 : 1,
+      this.audioContext.currentTime,
+    );
   }
 
   public setVolume(volume: number): void {
-    this.currentAudio.forEach(({ gainNode, originalVolume }) => {
-      // eslint-disable-next-line no-param-reassign
-      gainNode.gain.value = originalVolume * volume;
-    });
+    const clampedVolume = Math.max(0, Math.min(1, volume));
+
+    this.clientSetVolume = clampedVolume;
+    if (!this.audioContext || !this.clientVolumeGainNode) {
+      return;
+    }
+    this.clientVolumeGainNode.gain.setValueAtTime(
+      this.clientSetVolume,
+      this.audioContext.currentTime,
+    );
   }
 }